change input_x type int32 -> float32

+++ hello_cnn/cnn.py
@@ -26,7 +26,9 @@ class Cnn(object):
 
         # Placeholders for input, output and dropout
         self.input_x = tf.placeholder(
-            tf.int32, [None, sequence_length, embedding_size], name="input_x")
+            tf.float32,
+            [None, sequence_length, embedding_size],
+            name="input_x")
         self.input_y = tf.placeholder(
             tf.float32, [None, num_classes], name="input_y")
         self.dropout_keep_prob = tf.placeholder(
